+++ src/manage.h
@@ -124,7 +124,7 @@ typedef struct {
   gnss_signal_t sid;      /**< Signal identifier. */
   u32 sample_count;       /**< Reference NAP sample count. */
   float carrier_freq;     /**< Carrier frequency Doppler (Hz). */
-  float code_phase;       /**< Code phase (chips). */
+  double code_phase;      /**< Code phase (chips). */
   u32 chips_to_correlate; /**< Chips to integrate over. */
   float cn0_init;         /**< C/N0 estimate (dBHz). */
   s8 elevation;           /**< Elevation (deg). */
